Added sampleMLA.html and formatting scripts
I built the core editor and the parts that go along with it.

Using the sample of an MLA paper, I constructed and isolated the code to
write/format an MLA paper so that I can customize it with other formats.

-sampleMLA.html
Although called MLA, the only part of this page related to MLA are the
hard-coded links to the script.

formats\
This folder contains the various types of formats that can be
crowd-sourced and uploaded by users.

input.js
Uses simple rules to create the editor interface

build.js
Uses simple rules to convert the UI into a properly formatted paper/xml
(Although this uses broad rules)

mla\format.js
This contains not the infrastructure, but the front end to generating
the edtior interface and interpreting the format into a paper.