Improved Paragraphing and started Pagination
sampleMLA, build.js, input.js,and format.js all received updates that
enable them to write out input to a JSON object, then build that object
into a preview that includes proper paragraphs with tabs and
single/double spacing.

pagination.html has been started to generate print-ready documents that
have proper and adjustable margins and page sizes. (This would include
proper headers and footers.) So far, it is very 'in-progress'.

Since I'm now building this in NetBeans, a few project files are in the
nbproject\ folder. These can be ignored.