Reciting
The biggest change is allowing you to cite a source multiple times in
the document by matching titles. However, a survey of people shows that
this title match will need to be included by secondary authors and
editions.

LongQuotes have been added although this leads to a bug in formatting.
This is still in need of a fix.

The Rangy library was added to get and place the cursor, although it has
yet to be integrated.

Smarter bibliography only adds sources to the bibliography if that
source is cited in your document.